Add props validation with length
