+++ src/components/Header.js
@@ -1,10 +1,13 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
-const Header = ({ title }) => {
-  return (
-    <div>
-      <header>{title}</header>
-    </div>
-  );
+const Header = ({ title }) => (
+  <div>
+    <header>{title}</header>
+  </div>
+);
+Header.propTypes = {
+  title: PropTypes.string.isRequired,
 };
+
 export default Header;
